@@ -2,8 +2,8 @@ import 'dart:async';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:flutter/material.dart';
 import 'package:camera/camera.dart';
-import 'package:flutter_vibrate/flutter_vibrate.dart';
 import 'package:wms_app/pages/loadingPage.dart';
+import 'package:wms_app/utils.dart';
 
 // page which has future builder contraint should probably be called content
 // the loagingage loadingcontent
@@ -11,73 +11,64 @@ import 'package:wms_app/pages/loadingPage.dart';
 
 // this shoudld pe put into the dependency injector
 class CameraViewController {
+  static late CameraViewController instance = CameraViewController();
+
   // rename to 'scanning controll'?
-  static CameraImage streamImage;
+  static CameraImage? streamImage;
   //static XFile photoFile;
   static void updateCurrentImage(CameraImage img) {
     streamImage = img;
-
     //print("imageStream");
   }
 
-  static void pauseImageStream() {
-    _cameraController.stopImageStream();
+  static void pauseImageStream() async {
+    (await _cameraController).stopImageStream();
   }
 
-  static void resumeImageStream() {
-    _cameraController.startImageStream(updateCurrentImage);
+  static void startImageStream() async {
+    (await _cameraController).startImageStream(updateCurrentImage);
   }
 
-  static CameraController _cameraController;
-  static Future<CameraController> getCameraControllerInstance() async {
-    if (_cameraController == null) {
-      var cameras = await availableCameras();
-      if (cameras?.length == 0) {
-        print("the device did not have a camera");
-        return null;
-      }
-      _cameraController = CameraController(cameras[0], ResolutionPreset.max);
-      await _cameraController.initialize();
-      //_cameraController.setFlashMode(FlashMode.always); // null crash...
+  static late Future<CameraController> _cameraController =
+      _createCameraController();
 
-      // use stream image, or when commented out take ordinary photo to file and use
-      _cameraController.startImageStream(updateCurrentImage);
+  static Future<CameraController> _createCameraController() async {
+    var cameras = await availableCameras();
+    if (cameras.length == 0) {
+      throw new Exception("You need a device with camera use WMS App");
     }
-    // need a new controller each time rerendered
-
-    return _cameraController;
+    var cameraController = CameraController(cameras[0], ResolutionPreset.max);
+    await cameraController.initialize();
+    return cameraController;
   }
 
-  static AudioCache _audioCache;
-  static bool _canVibrate;
+  static AudioCache? _audioCache;
+  // static bool? _canVibrate;
   static Future<void> scanningSuccessfull() async {
     if (_audioCache == null) {
       _audioCache = AudioCache();
       // low latency mode allows fast and immedite beeps: https://stackoverflow.com/questions/59610504/flutter-audioplayers-delay
-      _audioCache.fixedPlayer = AudioPlayer(mode: PlayerMode.LOW_LATENCY);
-      _audioCache.fixedPlayer.setVolume(0);
-    }
-
-    if (_canVibrate == null) {
-      _canVibrate = await Vibrate.canVibrate;
+      _audioCache?.fixedPlayer = AudioPlayer(mode: PlayerMode.LOW_LATENCY);
+      _audioCache?.fixedPlayer?.setVolume(0);
     }
 
     /*
-    if (_canVibrate) {
-      Vibrate.feedback(FeedbackType.success);
-    } // vibration is made first despite called, but feels ok
+    // required compatible package, also feature might noe bed needed
+    if (_canVibrate == null) {
+      _canVibrate = await Vibrate.canVibrate;
+    }
     */
 
     _audioCache
-        .play("sounds/scanner_beep.mp3"); // (should be able to use waw also)
+        ?.play("sounds/scanner_beep.mp3"); // (should be able to use waw also)
   }
 
   static Future<XFile> takePhoto() async =>
-      (await getCameraControllerInstance()).takePicture();
+      (await _cameraController).takePicture();
 }
 
 class CameraView extends StatefulWidget {
-  Size size;
+  Size? size;
   CameraView([this.size]);
   @override
   _State createState() => _State();
@@ -86,10 +77,10 @@ class CameraView extends StatefulWidget {
 // WidgetsBindingObserver
 // needed to to detect app lifecycle events: https://medium.com/pharos-production/flutter-app-lifecycle-4b0ab4a4211a
 class _State extends State<CameraView> {
-  FutureBuilder futureBuilder() => FutureBuilder<CameraController>(
-      future: CameraViewController.getCameraControllerInstance(), // <--!!!
+  FutureBuilder futureBuilder() => FutureBuilder<CameraController?>(
+      future: CameraViewController._cameraController, // <--!!!
       builder:
-          (BuildContext context, AsyncSnapshot<CameraController> snapshot) {
+          (BuildContext context, AsyncSnapshot<CameraController?> snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           // https://stackoverflow.com/questions/52847534/flutter-futurebuilder-returning-null-error-triggered
           return LoadingPage();
@@ -98,7 +89,7 @@ class _State extends State<CameraView> {
         return content(snapshot.data);
       });
 
-  Widget content(CameraController controller) {
+  Widget content(CameraController? controller) {
     if (controller == null) {
       return Container(
           child: Center(
